requirement.txt
1.Tkinter
2.matplotlib
3.PIL